+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [1.2.0] - 2022-12-05
 ### Added
 - Print the soonest due date of the users' loans in the table. This helps to find the date you have to go to the library.
+- Add command to renew loans, with `loans renew`. The new loans will be printed after the renewal.
 
 ## [1.1.0] - 2022-11-27
 ### Added